A cryptocurrency price tracker to monitors the current value of digital currencies and predicting the future value of currency.
A cryptocurrency price tracker monitors the current value of digital currencies, in addition to information about market capitalization, trading volume, price charts, and other types of data.

The tracker's accuracy can have a significant bearing on an investor's decisions; because of this, investors should choose their price tracker carefully.

p.s. : I do not recommend people to use this project to use in business work but to use in understanding how prediction works and learn from it.